ENH: Add pluggable markups architecture
- This adds a singleton class to handle the registration of new markups.
This class can be used to both register and check registered markups and
associated widgets. This class supports pluggable markups.

- This adds a new factory function to vtkSlicerWidgets. In the context of
enabling pluggable markups, this is needed to allow the displayable
manager to create the right type of widget for a give markup node. The
types of created widgets are then resolved in run-time.

- This removes the hard-coded markups from
vtkMRMLMarkupsDisplayableManager and adapts the DM to a pluggable
architecture where the markups are registered using
vtkMRMLMarkupsRegistrationFactory.

- This uses the new markups pluggable archictecture to register the
markups defined in the Markups module.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>