ENH: Extend vtkMRMLMarkups nodes
Some of the components supporting markups have hard-coded variables that
represents genuine attributes of Markups:

  - Placement icon
  - Markup name
  - Markup short name

Here, vtkMRMLMarkups nodes are extended with these attributes. This is
useful in the context of pluggable markups, where we need to keep track
of the registered markups and all associated pieces of information
associated to these markups that don't belong to the markup node
classes. By extending the classes with these attributes, we avoid
keeping structures (e.g., maps) to keep track of relationships (e.g.,
markup and placement icon).