ENH: Bring markups naming consistency
This extends the markups labels naming by adding a short-name version of
the labels. This short-name version of the labeling is established as
default. Disambiguation of open and close curves is achieved by setting
the label to 'OC' and 'CC' respectively.

The reason for this change is to bring consistency in the naming of
markups. Before this commit, placing markups from the selection node
toolbar will use short naming, while placing markups from the Markups
module GUI would use long names.

This closes Slicer/Slicer#5196.